Fix sofia launcher logging
