+++ hydra_plugins/sofia_launcher/launcher.py
@@ -11,8 +11,9 @@ from omegaconf import DictConfig
 from omegaconf import open_dict
 from typing import Sequence, Optional
 from hydra.plugins.launcher import Launcher
+from hydra.core.hydra_config import HydraConfig
 from hydra.types import TaskFunction, HydraContext
-from hydra.core.utils import JobReturn, run_job, setup_globals
+from hydra.core.utils import JobReturn, run_job, setup_globals, configure_log
 
 from typing import Callable
 from posixpath import join as urljoin
@@ -20,43 +21,13 @@ from gymnos.services.sofia import SOFIA
 from gymnos.cli.utils import print_config, find_trainer_dependencies
 
 from .config import Device
-from .utils import get_current_revision, print_launcher, print_dependencies
-
-logger = logging.getLogger(__name__)
-
-
-class SOFIAProjectNotFound(Exception):
-
-    def __init__(self, project_name):
-        message = f"Project `{project_name}` not found"
-        super().__init__(message)
-
-
-def launch(args: Sequence[str], project_name: str, ref: str, device: Device) -> Callable:
-    def entrypoint(config):
-        response = SOFIA.create_project_job(args, project_name, ref, device.value)
-
-        if not response.ok:
-            if response.status_code == 404:
-                raise SOFIAProjectNotFound(project_name)
-            response.raise_for_status()
-
-        data = response.json()
-
-        job_url = urljoin(SOFIA.domain, data["project"]["user"]["username"], "projects", project_name,
-                          "jobs", data["name"])
-        logging.info(f"Project job successfully created at {job_url}")
-
-    return entrypoint
+from .core import launch
+from .utils import get_current_revision
 
 
 class SOFIALauncher(Launcher):
 
     def __init__(self, project_name: str, ref: str = None, device: Device = "CPU", show=True):
-        if ref is None:
-            logger.info("Revision not specified. Trying to retrieve the current revision for working directory")
-            ref = get_current_revision()
-
         self.project_name = project_name
         self.ref = ref
         self.device = device
@@ -72,34 +43,14 @@ class SOFIALauncher(Launcher):
         self.task_function = task_function
 
     def launch(self, job_overrides: Sequence[Sequence[str]], initial_job_idx: int) -> Sequence[JobReturn]:
-        job_returns = []
-
+        # Configure launcher
         setup_globals()
+        configure_log(self.config.hydra.job_logging, self.config.hydra.verbose)
 
-        if self.show:
-            print_launcher(self.config.hydra.launcher)
+        logger = logging.getLogger(__name__)
 
-        for idx, args in enumerate(job_overrides):
-            sweep_config = self.hydra_context.config_loader.load_sweep_config(self.config, list(args))
+        if self.ref is None:
+            logger.info("Revision not specified. Trying to retrieve the current revision for working directory")
+            self.ref = get_current_revision()
 
-            with open_dict(sweep_config):
-                sweep_config.hydra.job.id = str(uuid.uuid4())
-                sweep_config.hydra.job.num = initial_job_idx + idx
-
-            if sweep_config.show_config:
-                print_config(sweep_config, ("trainer", "data", "test"))
-
-            if sweep_config.show_dependencies:
-                dependencies = find_trainer_dependencies(sweep_config.trainer)
-                print_dependencies(dependencies)
-
-            job_return = run_job(
-                launch(args, self.project_name, self.ref, self.device),
-                config=sweep_config,
-                hydra_context=self.hydra_context,
-                job_dir_key="hydra.sweep.dir",
-                job_subdir_key="hydra.sweep.subdir",
-            )
-            job_returns.append(job_return)
-
-        return job_returns
+        return launch(self, job_overrides, initial_job_idx)
